Refactors dns code to it's own package
There's more work to be done making this dns code nicer.
But this is a good start, making main.go easily understandable as an entry point.

Error handling, util-type stuff can and should be factored out further.
